moved tests for item to itemservice test from itemservicetests
