+++ Bakendi/ThjonustukerfiTests/Tests/Order/OrderServiceTests.cs
@@ -180,31 +180,5 @@ namespace ThjonustukerfiTests.Tests
                 }
             };
         }
-
-        [TestMethod]
-        public void SearchItem_should_return_a_ItemStateDTO()
-        {
-            //* Arrange
-            var retDTO = new ItemStateDTO
-            {
-                Id = 1,
-                OrderId = 1,
-                Type = "Test",
-                State = "Í vinnslu",
-                DateModified = DateTime.Now
-            };
-            // Mock the method
-            _orderRepoMock.Setup(method => method.SearchItem("someString")).Returns(retDTO);
-
-            // Create controller
-            _orderService = new OrderService(_orderRepoMock.Object, _customerRepoMock.Object);
-
-            //* Act
-            var response = _orderService.SearchItem("someString");
-
-            //* Assert
-            Assert.IsNotNull(response);
-            Assert.IsInstanceOfType(response, typeof(ItemStateDTO));
-        }
     }
 }
